// src/app/models/Permission.ts
export const enum Permission {
  ADMINISTER_ROLES = "ADMINISTER_ROLES",
  ASSIGN_ISSUES = "ASSIGN_ISSUES",
  CREATE_ISSUES = "CREATE_ISSUES",
  CANCEL_ISSUES = "CANCEL_ISSUES",
  EDIT_ISSUES = "EDIT_ISSUES",
  DELETE_ISSUES = "DELETE_ISSUES",
  UPDATE_ISSUES_STATUS = "UPDATE_ISSUES_STATUS",
  MARK_ISSUES_FOR_CLOSURE = "MARK_ISSUES_FOR_CLOSURE",
  CLOSE_ISSUES = "CLOSE_ISSUES",
  VIEW_WATCHERS = "VIEW_WATCHERS",
  MANAGE_WATCHERS = "MANAGE_WATCHERS",
  ADD_COMMENTS = "ADD_COMMENTS",
  DELETE_ALL_COMMENTS = "DELETE_ALL_COMMENTS",
  DELETE_OWN_COMMENTS = "DELETE_OWN_COMMENTS",
  EDIT_ALL_COMMENTS = "EDIT_ALL_COMMENTS",
  EDIT_OWN_COMMENTS = "EDIT_OWN_COMMENTS",
}